requirements.txt
streamlit==1.13.0
tensorflow-cpu==2.10.0
Pillow==9.2.0
numpy==1.21.6
pandas==1.3.5
